+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Size of data packets is limited to 65535 bytes.
+- Update interval is now expressed as centiseconds, in accordance with the babel
+  RFC.
 
 ### Fixed